Modificacion del panel de admin separando el listado de productos
