+++ src/components/PanelAdministrador/Admin.module.css
@@ -1,85 +1,18 @@
 .panel {
   padding: 20px;
+  font-size: 30px;
   align-items: center;
-  text-align: right;
-}
-
-
-h2 {
-  font-size: 40px;
-  text-align: center;
-  color: #549ad9;
-
-}
-
-
-
-.actions {
-  display: flex;
-  flex-direction: column;
-  text-align: right;
-  margin-left: 20px;
- 
-}
-
-.acciones{
-  display: flex;
-  width: fit-content;
-}
-
-.adminTable {
-
-width: 900px;
-
-}
-
-thead {
-  width: 900px;
-  color: #549ad9;
-
- 
-}
-
-.name{
- 
-}
-
-.columnas{
   text-align: left;
-  
+  height: 400px;
 }
 
-tbody {
-  display: flex;
-  flex-direction: column;
-  gap: 5px;
-  font-size: 12px;
- 
-}
 
-tr {
-  display: flex;
-  align-items: center;
-  text-align: center;
-  margin-left: 20px;
-  justify-content: space-between;
-}
-
-td {
-  justify-content: space-between;
-  gap: 20px;
-
- 
-
-}
 
 button {
   padding: 10px;
   border-radius: 5px;
   cursor: pointer;
   outline: none;
+
 }
 
-.save{
-  color: red;
-}
