+++ src/components/PanelAdministrador/Inventory.module.css
@@ -1,0 +1,62 @@
+.panel {
+  padding: 20px;
+  align-items: center;
+  text-align: right;
+  height: 500px;
+}
+
+h2 {
+  font-size: 30px;
+  text-align: center;
+  color: #549ad9;
+}
+
+.actions {
+  display: flex;
+  flex-direction: column;
+  text-align: right;
+  margin-left: 20px;
+}
+
+
+.adminTable {
+  width: 900px;
+}
+
+thead {
+  width: 900px;
+  color: #549ad9;
+}
+
+.columnas {
+  text-align: left;
+
+}
+
+tbody {
+  display: flex;
+  flex-direction: column;
+  gap: 2px;
+  font-size: 12px;
+}
+
+tr {
+  display: flex;
+  align-items: center;
+  text-align: center;
+  margin-left: 20px;
+  justify-content: space-between;
+}
+
+td {
+  justify-content: space-between;
+  gap: 20px;
+}
+
+button {
+  padding: 10px;
+  border-radius: 5px;
+  cursor: pointer;
+  outline: none;
+ 
+}
