+++ src/components/PanelAdministrador/RegistrarProducto.module.css
@@ -7,6 +7,7 @@
   border: 1px solid #ccc;
   border-radius: 5px;
   background-color: #f5e9fc;
+  margin-bottom: 30px;
 }
 
 .form {
